feat: split state plugin from runtime package
