+++ packages/runtime/plugin-runtime/package.json
@@ -215,12 +215,6 @@
     "@loadable/babel-plugin": "5.15.3",
     "@loadable/component": "5.15.3",
     "@loadable/server": "5.15.3",
-    "@modern-js-reduck/plugin-auto-actions": "^1.1.10",
-    "@modern-js-reduck/plugin-devtools": "^1.1.10",
-    "@modern-js-reduck/plugin-effects": "^1.1.10",
-    "@modern-js-reduck/plugin-immutable": "^1.1.10",
-    "@modern-js-reduck/react": "^1.1.10",
-    "@modern-js-reduck/store": "^1.1.10",
     "@modern-js/plugin": "workspace:*",
     "@modern-js/plugin-data-loader": "workspace:*",
     "@modern-js/plugin-v2": "workspace:*",
