+++ packages/runtime/plugin-runtime/src/common.ts
@@ -1,19 +1,15 @@
 import type { Plugin, RuntimePluginFuture } from './core/plugin';
 import type { RouterConfig } from './router';
-import type { StateConfig } from './state';
 
 export const isBrowser = () =>
   typeof window !== 'undefined' && window.name !== 'nodejs';
 
 export interface AppConfig {
-  state?: StateConfig | boolean;
   router?: Pick<RouterConfig, 'future' | 'basename'>;
   [key: string]: any;
 }
 
 export interface RuntimeConfig {
-  state?: StateConfig;
-  stateByEntries?: { [name: string]: StateConfig };
   plugins?: (Plugin | RuntimePluginFuture)[];
   [key: string]: any;
 }
