+++ packages/runtime/plugin-runtime/src/config.ts
@@ -1,9 +1,7 @@
 import type { RouterConfig } from './router';
-import type { StateConfig } from './state';
 
 export interface RuntimeUserConfig {
   runtime?: {
-    state?: boolean | StateConfig;
     router?: boolean | RouterConfig;
   };
 }
