--- packages/runtime/plugin-runtime/tests/state/index.test.ts
@@ -1,9 +1,0 @@
-import plugin from '../../src/state';
-import { statePlugin } from '../../src/state/cli';
-
-describe('plugin-state', () => {
-  it('default', () => {
-    expect(plugin).toBeDefined();
-    expect(statePlugin).toBeDefined();
-  });
-});
