+++ packages/runtime/plugin-state/.npmignore
@@ -1,0 +1,31 @@
+.DS_Store
+
+.pnp
+.pnp.js
+.env.local
+.env.*.local
+*.log*
+
+node_modules/
+*.tsbuildinfo
+.eslintcache
+
+coverage/
+output/
+output_resource/
+tests/
+
+.vscode/**/*
+!.vscode/settings.json
+!.vscode/extensions.json
+.idea/
+
+src/
+
+modern.config.*
+jest.config.js
+.eslintrc.js
+.eslintrc
+tsconfig.json
+CHANGELOG.md
+
