+++ packages/runtime/plugin-state/jest.config.js
@@ -1,0 +1,7 @@
+const sharedConfig = require('@scripts/jest-config');
+
+/** @type {import('@jest/types').Config.InitialOptions} */
+module.exports = {
+  ...sharedConfig,
+  rootDir: __dirname,
+};
