+++ packages/runtime/plugin-state/modern.config.js
@@ -1,0 +1,5 @@
+const { universalBuildConfig } = require('@scripts/build');
+
+module.exports = {
+  buildConfig: universalBuildConfig,
+};
