+++ packages/runtime/plugin-state/package.json
@@ -1,0 +1,104 @@
+{
+  "name": "@modern-js/plugin-state",
+  "description": "A Progressive React Framework for modern web development.",
+  "homepage": "https://modernjs.dev",
+  "bugs": "https://github.com/web-infra-dev/modern.js/issues",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/web-infra-dev/modern.js",
+    "directory": "packages/runtime/plugin-router-v5"
+  },
+  "license": "MIT",
+  "keywords": [
+    "react",
+    "framework",
+    "modern",
+    "modern.js"
+  ],
+  "version": "2.65.4",
+  "jsnext:source": "./src/index.ts",
+  "types": "./src/cli/index.ts",
+  "main": "./dist/cjs/cli/index.js",
+  "module": "./dist/esm/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/types/cli/index.d.ts",
+      "jsnext:source": "./src/cli/index.ts",
+      "default": "./dist/cjs/cli/index.js"
+    },
+    "./runtime": {
+      "types": "./dist/types/runtime/index.d.ts",
+      "jsnext:source": "./src/runtime/index.ts",
+      "default": "./dist/esm/runtime/index.js"
+    },
+    "./cli": {
+      "types": "./dist/types/cli/index.d.ts",
+      "jsnext:source": "./src/cli/index.ts",
+      "default": "./dist/cjs/cli/index.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      ".": [
+        "./dist/types/cli/index.d.ts"
+      ],
+      "types": [
+        "./types/index.d.ts"
+      ],
+      "cli": [
+        "./dist/types/cli/index.d.ts"
+      ],
+      "runtime": [
+        "./dist/types/runtime/index.d.ts"
+      ]
+    }
+  },
+  "scripts": {
+    "dev": "modern-lib build --watch",
+    "prepublishOnly": "only-allow-pnpm",
+    "new": "modern-lib new",
+    "build": "modern-lib build",
+    "test": "jest --passWithNoTests"
+  },
+  "dependencies": {
+    "@modern-js/types": "workspace:*",
+    "@modern-js/runtime-utils": "workspace:*",
+    "@modern-js/utils": "workspace:*",
+    "@modern-js-reduck/plugin-auto-actions": "^1.1.10",
+    "@modern-js-reduck/plugin-devtools": "^1.1.10",
+    "@modern-js-reduck/plugin-effects": "^1.1.10",
+    "@modern-js-reduck/plugin-immutable": "^1.1.10",
+    "@modern-js-reduck/react": "^1.1.10",
+    "@modern-js-reduck/store": "^1.1.10",
+    "@swc/helpers": "0.5.13"
+  },
+  "peerDependencies": {
+    "react": ">=17",
+    "react-dom": ">=17",
+    "@modern-js/runtime": "workspace:^2.65.4"
+  },
+  "devDependencies": {
+    "@modern-js/app-tools": "workspace:*",
+    "@modern-js/runtime": "workspace:*",
+    "@scripts/build": "workspace:*",
+    "@scripts/jest-config": "workspace:*",
+    "@testing-library/react": "^13.4.0",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.1",
+    "@types/invariant": "^2.2.30",
+    "@types/jest": "^29",
+    "@types/node": "^14",
+    "jest": "^29",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5"
+  },
+  "sideEffects": false,
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/",
+    "access": "public",
+    "provenance": true,
+    "types": "./dist/types/cli/index.d.ts"
+  }
+}
