+++ packages/runtime/plugin-state/src/cli/types.ts
@@ -1,0 +1,7 @@
+import type { StateConfig } from '../runtime/plugin';
+
+declare module '@modern-js/app-tools' {
+  export interface RuntimeUserConfig {
+    state?: StateConfig | boolean;
+  }
+}
