+++ packages/runtime/plugin-state/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@modern-js/tsconfig/base",
+  "compilerOptions": {
+    "declaration": false,
+    "jsx": "preserve",
+    "baseUrl": "./",
+    "isolatedModules": true,
+    "paths": {}
+  },
+  "include": ["src", "types"]
+}
