+++ packages/runtime/plugin-state/type.d.ts
@@ -1,0 +1,16 @@
+import type { StateConfig } from './dist/types';
+
+declare module '@modern-js/runtime' {
+  interface AppConfig {
+    state?: StateConfig;
+  }
+
+  interface RuntimeConfig {
+    state?: StateConfig;
+  }
+}
+declare module '@modern-js/app-tools' {
+  interface RuntimeUserConfig {
+    state?: StateConfig;
+  }
+}
